BoostedDarkMatter

Documentation for BoostedDarkMatter.

Kinematics

BoostedDarkMatter.Kinematics — Module

The Kinematics sub-module contains kinematics function for elastic two body scattering,

\[p_1 + p_2 → p_3 + p_4.\]

The convention is $1 → 3$ and $2 → 4$. Currently, the functions in this module are given for the laboratory frame where the particle 2 is at rest.

BoostedDarkMatter.Kinematics.T1_min — Method

T1_min(T4, m1, m2)

Minimal kinetic energy of particle 1 required to obtain a recoil energy of T4 for the stationary target 2 in elastic scattering,

\[\begin{aligned} T_1^{\min} &= \left(\frac{T_4}{2} - m_2\right) \left(1 ± \sqrt{1 + \frac{2T_4}{m_1}\frac{(m_1+m_2)^2}{(2m_2-T_4)^2}}\right)\\ &= \frac{T_4}{2} - m_2 + \sqrt{\left(\frac{T_4}{2} - m_2\right)^2 + \frac{T_4(m_1+m_2)^2}{2m_1}}. \end{aligned}\]

In the case of $T_4 < 2m_2$ and $\left(\frac{T_4}{2} - m_2\right)^2 ≫ \frac{T_4(m_1+m_2)^2}{2m_1}$, we need to expand the square root to avoid precision loss, and the result is

\[T_4^{\min} ≈ \frac{T_4(m_1+m_2)^2}{2m_1(2m_2 - T_4)}.\]

Arguments

T4: Recoil kinetic energy of the target particle 2.

m1: Mass of the incident particle 1.

m2: Mass of the target particle 2.

BoostedDarkMatter.Kinematics.T1_min_nonrel — Method

T1_min_nonrel(T4, m1, m2)

Non-relativistic version of T1_min,

\[ T_1^{\min} = \frac{m_1 m_2 T_4}{4 μ_{12}^2} = \frac{(m_1 + m_2)^2 T_4}{4m_1 m_2}.\]

BoostedDarkMatter.Kinematics.T4_max — Method

T4_max(T1, m1, m2)

Maximum recoil energy of the stationary target particle 2 in elastic scattering,

\[T_4^{\max} = \frac{2m_2 T_1 (T_1 + 2m_1)}{2m_2 T_1 + (m_1 + m_2)^2}.\]

Arguments

T1: Kinetic energy of the incident particle 1.

m1: Mass of the incident particle 1.

m2: Mass of the target particle 2.

BoostedDarkMatter.Kinematics.T4_max_nonrel — Method

T4_max_nonrel(T1, m1, m2)

Non-relativistic version of T4_max,

\[T_4^{\max} = \frac{4 μ_{12} T_1}{m_1 m_2} = \frac{4 m_1 m_2 T_1}{(m_1 + m_2)^2}.\]

BoostedDarkMatter.Kinematics.reduce_m — Method

reduce_m(m1, m2)

Reduced mass $μ_{12} = \frac{m_1 m_2}{m_1 + m_2}$.

Index

BoostedDarkMatter.Kinematics

BoostedDarkMatter.Kinematics.T1_min

BoostedDarkMatter.Kinematics.T1_min_nonrel

BoostedDarkMatter.Kinematics.T4_max

BoostedDarkMatter.Kinematics.T4_max_nonrel

BoostedDarkMatter.Kinematics.reduce_m