BoostedDarkMatter

Documentation for BoostedDarkMatter.

BoostedDarkMatter.CRDM — Type

struct CRDM{T <: Number} <: BDM

Cosmic ray boosted dark matter (CRDM) model.

Fields

kfactor::Dict: Dictionary of the $K$ factor as a function of total kinetic energy for each CR species. See Ref.[3] for definition.

crdist::Dict: Dictionary of CR distributions. The CR distributions have to be callable with the signature (r, b, l, Ek), where (r, b, l) is the galactic coordinate and Ek is the kinetic energy of the CR particle.

Ekn_max::Dict: Dictionary of the maximal kinetic energy per nucleon for CR species.

dmprofile::DMProfile: DM profile.

xsec::XSec: DM-CR scattering cross section.

selected_cr::Vector{String}: The selected CR species for calculation.

rsun::T<:Number: The distance between the Sun and the Galactic Center.

References

[1] Bringmann, T., Pospelov, M., 2019. Novel Direct Detection Constraints on Light Dark Matter. Phys. Rev. Lett. 122, 171801.

[2] Ema, Y., Sala, F., Sato, R., 2018. Light Dark Matter at Neutrino Experiments. Phys. Rev. Lett. 122, 181802.

[3] Xia, C., Xu, Y.-H., Zhou, Y.-F., 2022. Production and attenuation of cosmic-ray boosted dark matter. JCAP 02, 028.

BoostedDarkMatter.CRFluxLISElectron — Type

CRFluxLISElectron{T <: Number, U <: Number} <: CRFluxElectron

A parameterization of local CR electron flux:

\[ F(T) = \begin{cases} \frac{1.181×10^{11} T^{-12.061}}{1 + 4.307 × 10^8 T^{-9.269} + 3.125×10^8 T^{-10.697}}, & T < 6.88 \text{GeV} \\ 995.598 T^{-3.505} + 4.423 T^{-2.620}, & T ≥ 6.88 \text{GeV} \end{cases},\]

where the units are m$^{-2}$sr$^{-1}$s$^{-1}$GeV$^{-1}$. The parameterization is valid from $E_\text{cut1} = 2$ MeV up to $E_\text{cut2} = 90$ GeV. Nevertheless, the two cutoff energies are set as free parameters for convenience.

Fields

Ecut1::T<:Number=0.002: Lower cutoff kinetic energy [GeV].

Ebr::T<:Number=6.88: Kinetic energy break [GeV].

Ecut2::T<:Number=90.0: Upper cutoff kinetic energy [GeV].

unitflux::U<:Number=1.0: Unit of the flux (for Unitful.jl support).

References

Boschini, M. J., et al. "HelMod in the works: from direct observations to the local interstellar spectrum of cosmic-ray electrons." The Astrophysical Journal 854.2 (2018): 94.

BoostedDarkMatter.CRFluxLISHelMod2017 — Type

CRFluxLISHelMod2017{T <: Number, U <: Number} <: CRFlux

A parametrization of local interstellar CR nuclei flux:

\[ F(R) × R^{2.7} = \begin{cases} ∑_{i=0}^5 a_i R^i, & R ≤ 1 \text{GV}, \\ b + \frac{c}{R} + \frac{d_1}{d_2 + R} + \frac{e_1}{e_2 + R} + \frac{f_1}{f_2 + R} + gR, & R ≥ 1 \text{GV} \end{cases},\]

where $R$ is rigidity and the unit is [m$^{-2}$ s$^{-1}$ sr$^{-1}$ GV$^{-1}$].

References

M. J. Boschini et al., 2017. Solution of Heliospheric Propagation: Unveiling the Local Interstellar Spectra of Cosmic-ray Species. Astrophys. J. 840, 115.

BoostedDarkMatter.CRFluxSBPLElectron — Type

CRFluxSBPLElectron{T <: Number, U <: Number, V <: Number} <: CRFluxElectron

Smooth broken power law parameterization of local CR electron flux:

\[ Φ_\text{CRE} = Φ_0 \left[ 1 + \left( \frac{E_\text{br1}}{E} \right)^k \right]^{(γ_1 - γ_2)/k} \left( \frac{E}{E_\text{br}} \right)^{-γ_2} \left[ 1 + \left( \frac{E}{E_\text{br2}} \right)^k \right]^{(γ_3 - γ_2)/k}.\]

Fields

phi0::T<:Number=5.02e-6: $Φ₀$ [m$^{-2}$sr$^{-1}$s$^{-1}$GeV$^{-1}$].

Ebr1::U<:Number=46.0: $E_\text{br1}$ [GeV].

Ebr2::U<:Number=987.8: $E_\text{br2}$ [GeV].

Ebr::U<:Number=300.0: $E_\text{br}$ [GeV].

k::V<:Number=10.0: $k$.

gamma1::V<:Number=3.24: $γ_1$.

gamma2::V<:Number=3.1: $γ_2$.

gamma3::V<:Number=3.89: $γ_3$.

References

Ding, Y. C., Li, N., Wei, C. C., Wu, Y. L., & Zhou, Y. F. (2021). Implications of a possible TeV break in the cosmic-ray electron and positron flux. Physical Review D, 103(11), 115010.

BoostedDarkMatter.Ek_to_v — Method

kinetic energy to velocity

BoostedDarkMatter.T0Tz — Method

T0Tz(atten::Attenuation, Tz, z; kwargs...)

Inverse of TzT0.

Parameters

atten: Attenuation object.

Tz: Kinetic energy at depth z.

z: The depth.

BoostedDarkMatter.TzT0 — Method

TzT0(atten::Attenuation, T0, z; kwargs...)

Parameters

atten: Attenuation object.

T0: Kinetic energy at the surface (z = 0).

z: The depth.

BoostedDarkMatter._los_bound — Method

Line of sight boundary.

BoostedDarkMatter._los_integrand — Method

Dark matter profile times cosmic ray flux.

BoostedDarkMatter.cache_dTdz! — Function

Cache dTdz

BoostedDarkMatter.crflux_Ekn — Method

crflux_Ekn(cr::CRFluxSBPLElectron, coordinates::Tuple, args...; kwargs...)
crflux_Ekn(cr::CRFlux, r, b, l, Ekn, args...; kwargs...)

Intensity in kinetic energy per nucleon $\frac{dI}{dE_\text{kn}}$ [m$^{-2}$s$^{-1}$sr$^{-1}$GeV$^{-1}$].

Arguments

T: kinetic energy of CR electron

BoostedDarkMatter.dTdz — Method

dTdz(atten::Attenuation, T; usecache=true, atol=0, rtol=1e-6, kwargs...)

Calculate $\frac{dT}{dz}$ at T with the attenuation model atten,

\[\frac{dT}{dz} = -\int_0^{T_\text{r}^{\max} n \frac{dσ}{dT_\text{r}} T_\text{r} dT_\text{r}.\]

Arguments

atten::Attenuation: The Attenuation object.

T: Kinetic energy.

Keywords

usecache: If true and the cache exists, then the cached _dTdz will be used.

atol=0: Absolute tolerence.

rtol=1e-6: Relative tolerence.

BoostedDarkMatter.dmflux — Method

dmflux(bdm::CRDM, Tchi; Ekn_cutoff=nothing, fluxunit=1.0, kwargs...)

Return the total flux of CRDM on the Earth.

Use the $K$ factor and the local cosmic ray spectrum to calculate the flux (per solid angle) of CRDM produced by selected CR species.

Arguments

bdm::CRDM: Cosmic ray boosted dark matter model.

Tchi: Kinetic energy of DM particle.

Keywords

Ekn_cutoff: Cutoff of CR kinetic energy per nucleon. If nothing, bdm.Ekn_max will be used (the default is nothing).

fluxunit=1.0: The unit of the flux.

kwargs...: Other keyword arguments are passed to the quad integrator.

BoostedDarkMatter.dxsecdT4 — Method

dxsecdT4(xsec::XSec, T4, T1, m1, m2, args...; kwargs...)

Differential cross section $\frac{dσ}{dT_4}(T_4, T_1, m_1, m_2)$.

Arguments

T4: Recoil energy of the target particle 2.

T1: Kinetic energy of the incident particle 1.

m1: Mass of the incident particle 1.

m2: Mass of the target particle 2.

BoostedDarkMatter.dxsecdT4 — Method

Warning: using non-relativistic approximation

BoostedDarkMatter.dxsecdT4 — Method

dxsecdT4(xsec::XSecPseudoScalarMediator, T4, T1, m1, m2, args...; kwargs...)

Pseudo scalar mediated fermionic DM-fermion scattering cross section.

BoostedDarkMatter.dxsecdT4 — Method

dxsecdT4(xsec::XSecScalarMediator, T4, T1, m1, m2, args...; kwargs...)

Scalar mediated fermionic DM-fermion scattering cross section.

BoostedDarkMatter.ff_helm — Method

ff_helm(Q2, A; s=units.fm)

Helm form factor:

\[ F_\text{Helm}(Q^2) = \frac{3 j_1(qR_1)}{qR_1} e^{-Q^2 s^2 / 2},\]

where $j_1$ is the spherical Bessel function of the first kind, $q = \sqrt{Q^2}$, $R_A = 1.2 A^{1/3}$ fm, $R_1 = \sqrt{R_A^2 - 5s^2}$, and $s = 1$ fm.

Arguments

Q2: Momentum transfer squared.

A: Number of nucleons.

Keywords

s=units.fm: The length scale, default to 1 fm.

References

Lewin, J.D.D., Smith, P.F.F., 1996. Review of mathematics, numerical factors, and corrections for dark matter experiments based on elastic nuclear recoil. Astroparticle Physics 6, 87–112.

Engel, J., 1991. Nuclear form factors for the scattering of weakly interacting massive particles. Physics Letters B 264, 114–119.

BoostedDarkMatter.flux_to_vdist — Method

convert flux to f(v)

BoostedDarkMatter.h5dump_bl_distribution — Method

Dump 2D (b, l) distribution to a hdf5 file.

BoostedDarkMatter.h5dumpflux — Method

Dump 3-D intensity to a HDF5 file.

BoostedDarkMatter.h5load_bl_distribution — Function

Load and interpolate the 2-D distribution dumped by h5dump_bl_distribution.

BoostedDarkMatter.h5loadflux — Function

Load and interpolate the 3-D intensity dumped by h5loadflux.

BoostedDarkMatter.kfactor! — Method

kfactor!(bdm::CRDM, Ti::AbstractVector; save_to_file=nothing, kwargs...)

Calculate and set the angular averaged effective distance.

Arguments

bdm::CRDM: For CRDM model.

Ti: Energy grid of CR kinetic energy.

Keywords

save_to_file=nothing: Switch whether save the results to a json file, default is no. Giving a file name means save to the file.

BoostedDarkMatter.kfactor! — Method

kfactor!(bdm::CRDM, kfactor::Number)
kfactor!(bdm::CRDM, kfactor::Function)
kfactor!(bdm::CRDM, jsonfile::String)

Set the angular averaged effective distance. Since the calculation of the $K$ factor is time-consuming, it is better to manually set it in advance.

The particles selected in the bdm.selected_cr field will be set.

Arguments

bdm::CRDM: For CRDM model.

kfactor: K factor function, given a single number means constant K factor.

BoostedDarkMatter.lorentz_gamma — Method

Lorentz gamma $1/\sqrt{1-v^2}$ with speed of light c = 1.

BoostedDarkMatter.nucleus_mass — Method

Mass from A.

BoostedDarkMatter.recoil_spectrum — Method

recoil_spectrum(xsec::XSec, T4, flux1, T1max, m1, m2, args...; attenuation=nothing)

Recoil spectrum of the target at rest in the scattering.

Arguments

xsec::XSec: Differential cross section.

T4: Recoil energy of the target particle 2.

flux1: Flux of the incident particle 1.

T1max: Maximal energy of flux1.

m1: Mass of the incident particle 1.

m2: Mass of the target particle 2.

BoostedDarkMatter.v_to_Ek — Method

velocity to kinetic energy

BoostedDarkMatter.vdist_to_flux — Method

Convert 1D velocity distribution f(v) to flux.

BoostedDarkMatter.vdist_to_flux_3d — Method

Convert 3D velocity distribution f(v, theta, phi) to flux.

Kinematics

BoostedDarkMatter.Kinematics — Module

The Kinematics sub-module contains kinematics function for elastic two body scattering,

\[p_1 + p_2 → p_3 + p_4.\]

The convention is $1 → 3$ and $2 → 4$. Currently, the functions in this module are given for the laboratory frame where the particle 2 is at rest.

BoostedDarkMatter.Kinematics.T1_min — Method

T1_min(T4, m1, m2)

Minimal kinetic energy of particle 1 required to produce a recoil energy T4 of target 2 at rest in elastic scattering,

\[\begin{aligned} T_1^{\min} &= \left(\frac{T_4}{2} - m_1\right) \left(1 ± \sqrt{1 + \frac{2T_4}{m_2}\frac{(m_1+m_2)^2}{(2m_1-T_4)^2}}\right)\\ &= \frac{T_4}{2} - m_1 + \sqrt{\left(\frac{T_4}{2} - m_1\right)^2 + \frac{T_4(m_1+m_2)^2}{2m_2}}. \end{aligned}\]

In the case of $T_4 < 2m_1$ and $\left(\frac{T_4}{2} - m_1\right)^2 ≫ \frac{T_4(m_1+m_2)^2}{2m_2}$, we need to expand the square root to avoid precision loss, and the result is

\[T_4^{\min} ≈ \frac{T_4(m_1+m_2)^2}{2m_2(2m_1 - T_4)}.\]

Here we simply use the function sqrtm1 provided by NumericalTools.jl to handle this case.

Arguments

T4: Recoil kinetic energy of the target particle 2.

m1: Mass of the incident particle 1.

m2: Mass of the target particle 2.

BoostedDarkMatter.Kinematics.T1_min_nonrel — Method

T1_min_nonrel(T4, m1, m2)

Non-relativistic version of T1_min,

\[ T_1^{\min} = \frac{m_1 m_2 T_4}{4 μ_{12}^2} = \frac{(m_1 + m_2)^2 T_4}{4m_1 m_2},\]

where $μ_{12} = m_1 m_2 / (m_1 + m_2)$ is the reduced mass.

BoostedDarkMatter.Kinematics.T1_min_q — Method

T1_min_q(q, ΔE, m1)

Minimal kinetic energy of particle 1 with given momentum transfer q, energy transfer ΔE, and mass m1. It is simply calculated from p1_min_q using the on-shell condition.

BoostedDarkMatter.Kinematics.T4_max — Method

T4_max(T1, m1, m2)

Maximum recoil energy of the stationary target particle 2 in elastic scattering,

\[T_4^{\max} = \frac{2m_2 T_1 (T_1 + 2m_1)}{2m_2 T_1 + (m_1 + m_2)^2}.\]

Arguments

T1: Kinetic energy of the incident particle 1.

m1: Mass of the incident particle 1.

m2: Mass of the target particle 2.

BoostedDarkMatter.Kinematics.T4_max_nonrel — Method

T4_max_nonrel(T1, m1, m2)

Non-relativistic version of T4_max,

\[T_4^{\max} = \frac{4 μ_{12} T_1}{m_1 m_2} = \frac{4 m_1 m_2 T_1}{(m_1 + m_2)^2}.\]

BoostedDarkMatter.Kinematics.p1_min_q — Method

p1_min_q(q, ΔE, m1)

Minimal momentum of particle 1 with given momentum transfer q and energy transfer ΔE,

\[ p_1^{\min} = \frac{q}{2} + \frac{ΔE}{2} \sqrt{1 + \frac{4m_1^2}{q^2 - ΔE^2}}\]

Note that the mass of the particle 1 m1 is assumed to be unchanged after the scattering, which guarantees $q > ΔE$.

Arguments

q: Momentum transfer.

ΔE: Energy transfer.

m1: Mass of the incident particle 1.

BoostedDarkMatter.Kinematics.reduce_m — Method

reduce_m(m1, m2)

Reduced mass $μ_{12} = \frac{m_1 m_2}{m_1 + m_2}$.

Index

BoostedDarkMatter.Kinematics

BoostedDarkMatter.CRDM

BoostedDarkMatter.CRFluxLISElectron

BoostedDarkMatter.CRFluxLISHelMod2017

BoostedDarkMatter.CRFluxSBPLElectron

BoostedDarkMatter.Ek_to_v

BoostedDarkMatter.Kinematics.T1_min

BoostedDarkMatter.Kinematics.T1_min_nonrel

BoostedDarkMatter.Kinematics.T1_min_q

BoostedDarkMatter.Kinematics.T4_max

BoostedDarkMatter.Kinematics.T4_max_nonrel

BoostedDarkMatter.Kinematics.p1_min_q

BoostedDarkMatter.Kinematics.reduce_m

BoostedDarkMatter.T0Tz

BoostedDarkMatter.TzT0

BoostedDarkMatter._los_bound

BoostedDarkMatter._los_integrand

BoostedDarkMatter.cache_dTdz!

BoostedDarkMatter.crflux_Ekn

BoostedDarkMatter.dTdz

BoostedDarkMatter.dmflux

BoostedDarkMatter.dxsecdT4

BoostedDarkMatter.dxsecdT4

BoostedDarkMatter.dxsecdT4

BoostedDarkMatter.dxsecdT4

BoostedDarkMatter.ff_helm

BoostedDarkMatter.flux_to_vdist

BoostedDarkMatter.h5dump_bl_distribution

BoostedDarkMatter.h5dumpflux

BoostedDarkMatter.h5load_bl_distribution

BoostedDarkMatter.h5loadflux

BoostedDarkMatter.kfactor!

BoostedDarkMatter.kfactor!

BoostedDarkMatter.lorentz_gamma

BoostedDarkMatter.nucleus_mass

BoostedDarkMatter.recoil_spectrum

BoostedDarkMatter.v_to_Ek

BoostedDarkMatter.vdist_to_flux

BoostedDarkMatter.vdist_to_flux_3d